+++ src/frontend/src/pages/MainPage/pages/assetsPage/components/dragWrapComponent/index.tsx
